优化资源上限逻辑

+++ Assets/Scripts/LZZ_Test/LevelChange/LevelInformationSO.cs
@@ -11,7 +11,7 @@ namespace ScriptableObjects
         //public int sceneId;
         public int levelID;
         //public float levelTime;
-        public float levelResource;
+        public int levelResource;
         public string levelName;
         public string levelIntro;
         public List<Image> enemies;
